Terraform to deploy DB and image using cloud run
This is almost working as a standalone automation, but the DB creation
within the instance still needs to be performed by hand. I will look
into using a provider to perform this step (and ideally the DDL steps
that the distributor performs in db.init so that the app doesn't need
schema modification permissions).

For now, this works if one has sufficient ACLs for the project and runs:
```bash
terraform init
terraform apply
```

This should bring up the DB and Distributor running in Cloud Run. To
create the Database, the easiest thing to do is to find the instance
within Cloud SQL in GCP UI, and create a database called "distributor".

After doing that, it's possible to query the distributor using curl:
```bash
curl -i -H \
"Authorization: Bearer $(gcloud auth print-identity-token)" \
$DISTRIBUTOR_URI/distributor/v0/logs
```

Note that you'll need to substitute the URI for the new instance that is
created on apply. This is output from `terraform apply` under the name
`distributor_uri` but you can find this in the Cloud Run area of GCP if
you lose it down the back of the sofa.

The messy secret URL is only while this is in experimental. When we have
this launched properly then it will have a public URL and not require
auth tokens.

Also TODO in a future PR before we can take this to prod is to add
integration to make the state be stored in a GCS bucket instead of
output into a local file.