Support distributor launching without witness file
This will need to be revisited, but doing this unblocks getting the distributor running in Cloud Run. In such an environment, there are no files. And even if we were doing this in GKE, we would probably look for a solution that didn't have a file embedded in the container.

The long-term solution to this problem will be to read the witnesses from a log, but we're not there yet.