Update terraform docs and script (#44)
The script was being flaky, failing to start cloud run because secrets were not available. Changed to provision all of the secrets and read access to them earlier so these can propagate while the rest of the script is running. This also makes it easier to read, as all of the secrets are in the same place.