Improve deployment experience and features
Two features are:
 1. Store terraform state in GCS buckets
 2. Support multiple deployments for different environments

Storing state in GCS buckets provides a canonical location for the
current deployment state, which better supports multiple maintainers.
Previously, the state was stored on the local developer machine. This
made life tedious and error prone for multiple maintainers, or even a
single maintainer with multiple machines.

Different environments allows us to avoid testing in prod, and to stage
releases properly. ci and prod are created for real world usage, and the
dev environment is an environment in which we can test out more risky
changes that won't be exposed to the outside world.

Note that these changes require terragrunt in addition to terraform.
This is a small open source tool that provides support for both of these
features. It breaks the chicken-and-egg loop of creating GCS buckets
using terraform, but needing a GCS bucket in which to store the state.
It also easily supports different live environments invoking the same
module with input parameters, allowing for clear code reuse while having
separate files for each environment. This is the least risky approach I
can see to supporting updates to one environment without modifying the
others.