Cloud build docker images for distributor (#83)
This adds terragrunt/terraform configurations that define cloud build triggers and steps. The action triggers on push to main in the github repo, and then builds a docker image, which is then pushed to artifact registry (which will also be created by the terraform config when applied, if it doesn't already exist). There are two versions: dev, which can be used by developers testing out changes to the build steps, and prod, which should only be updated using a pristine checkout of the main branch.

As part of this, the existing terragrunt configs for the serving stack were moved into a serving directory to form a clean separation between the cloudbuild configs and the serving configs. These can all be applied separately, which means that updating cloudbuild configs won't modify the serving stack.