Use docker images from artifact registry (#85)
This was using the images built by trillian-opensource-ci using the cloudbuild file deleted in this PR. This file is no longer needed after this change, as the image deployed will be the one built by the cloudbuild set up in the terragrunt configs in this repo. Coming changes will set up CI to be updated automatically when cloudbuild prod builds a new image.