Export metrics on main port (#86)
CloudRun doesn't allow services to expose multiple ports, thus it wasn't possible to scrape these metrics when exported on a different port. This changes the metrics to be exported on the default port, but the registration can be disabled via a flag in case we want to turn this off.