Terraform to set up slack integration
This required some setup first:
 - Creating a Slack app with webhooks: https://api.slack.com/apps/A06KYD43DPE/incoming-webhooks?
 - Setting up a secret in Secret Manager in the cloud project with the webhook URL

This will post to a new Slack channel in transparency-dev called `cloudbuild-dev`, which I've created as a private group with me as the only member for testing.

When we're happy with this, we can deploy a prod version which posts to a public channel in the Slack channel.