Override the Slack template with a custom version (#99)
The last version didn't actually work because the module didn't create the required json template. I've made some more changes to my fork that creates a default json template that can be overridden. We override that here with a custom template. This allows us to play around with the Slack template until we're happy with the notifications in dev, before promoting that to prod.