Tag docker images with git hashes and tags
The build step that previously existed now adds the short commit SHA as a docker tag, in addition to latest.

There is an additional trigger that fires on tags that will pull the docker image with the commit SHA corresponding to the tag, and then push it with the git tag as a docker tag.