Slack notifications are filtered to channels based on env
Builds triggered by the -dev trigger go to cloudbuild-dev, and ones from the -prod trigger go to the public channel. This cuts down on a lot of duplicate noise, and means that we can test in dev without creating concern that everything is broken on the public channel.

Note that the first parts of the filter are all the same as the default value for this predicate. The only diff is the trigger.endsWith.