Terraform: set witnesses per environment
With this I've removed the final use of the baked-in witnesses config. Witnesses must now be configured explicitly, either via flags or a config file.

The single witnesses.yaml file has been sharded into 3 files: one per environment. These files are the canonical description of witnesses allowed for each environment. The terraform script reads the environment-specific file and configures the flags appropriately.

This seems like a good place to leave the witness list until such time that we can resource the design and rollout of defining them in a log (or logs).