Expose an endpoint with all of the witness keys
Each of the environments will be configured differently, and thus using the single embedded config on both the client and the server to enumerate the witnesses was brittle/broken.

This change temporarily breaks the client until all distributors are updated to have the new endpoint, but we're still at v0 and I'll get this rolled out within a matter of hours so I'll avoid the tax of doing this the staged way. This is also a breaking change at the code level, but again, this is v0 and we can flex that licence.