Moved more config into common terragrunt file
Refactoring that moves the logic of parsing the witness file and transforming it into flags into the common file. Doing this meant I solved the problem of accessing local environment variables in the root include, which meant we could remove a second parsing of that file, and a duplicate definition of the environment. The env name now comes only from the directory name containing the leaf terragrunt file. Super nice.