Move dependabot to weekly
Also try to prevent it from bumping the minor version of Go until we explicitly want it to do that. I wanted to set this to > 1.22 but I have no idea how that will be interpreted. It could assume that 1.22.4 > 1.22 and thus keep us stuck. So I've picked an abnormally large patch version and hopefully that is clear.