Made gatekeeping of new checkpoints more permissive (#233)
This change has two main aspects, but they are related. The first is that we now overwrite a checkpoint from a witness even if we have previously seen one of the same size. This allows infrequently updated checkpoints to have fresh witness signature timestamps. The other change is to only check the witness note signature, and not the cosignatureV1. This is a temporary change to allow signatures from the fixed implementation of this signer (https://github.com/transparency-dev/formats/pull/153) to propagate into the distributor storage.