Switch back to using cosignature
This will reinstate timestamps in the witness signatures. This was dropped in #233 as a temporary measure to facilitate a transition between timestamp endianness. Any sigs that aren't verified are dropped, and thus the currently deployed distributors will drop the cosignature signatures.

If we want to support keeping both witness signatures around, we'll have to do some more engineering to have multiple verifiers per witness. We would also need to look at the merging code to ensure we only merge similar signatures. It seems easier to just drop the non-timestamped sigs on entry.